requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.21.0
dill==0.3.8
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
packaging==24.1
pandas==2.2.2
pillow==10.4.0
progressbar2==4.4.2
psutil==6.0.0
pyarrow==17.0.0
pyg-lib==0.4.0+pt24cu118
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.4.0+cu118
torch_cluster==1.6.3+pt24cu118
torch_geometric==2.5.3
torch_scatter==2.1.2+pt24cu118
torch_sparse==0.6.18+pt24cu118
torch_spline_conv==1.2.2+pt24cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
xxhash==3.5.0
yarl==1.9.4
